+++ firestore-palm-gen-text/functions/src/generator.ts
@@ -20,9 +20,9 @@ import {helpers, v1} from '@google-cloud/aiplatform';
 import * as logs from './logs';
 import {GoogleAuth} from 'google-auth-library';
 import {
-  APIGenerateTextRequest,
+  GLGenerateTextRequest,
   VertexPredictResponse,
-  APIGenerateTextResponse,
+  GLGenerateTextResponse,
 } from './types';
 import config from './config';
 
@@ -34,23 +34,13 @@ export interface TextGeneratorOptions {
   topK?: number;
   maxOutputTokens?: number;
   instruction?: string;
+  generativeSafetySettings?: GLGenerateTextRequest['safetySettings'];
 }
 
 export type TextGeneratorRequestOptions = Omit<
-  APIGenerateTextRequest,
+  GLGenerateTextRequest,
   'prompt' | 'model'
 >;
-
-type SafetyAttributes = {
-  blocked?: boolean;
-  scores?: number[];
-  categories?: string[];
-};
-
-export type TextGeneratorResponse = {
-  candidates: string[];
-  safetyAttributes?: SafetyAttributes;
-};
 
 export class TextGenerator {
   private generativeClient?: TextServiceClient;
@@ -64,6 +54,7 @@ export class TextGenerator {
   topP?: number;
   topK?: number;
   maxOutputTokens?: number;
+  generativeSafetySettings: TextGeneratorRequestOptions['safetySettings'];
 
   constructor(options: TextGeneratorOptions = {}) {
     this.temperature = options.temperature;
@@ -72,6 +63,7 @@ export class TextGenerator {
     this.maxOutputTokens = options.maxOutputTokens;
     this.candidateCount = options.candidateCount;
     this.instruction = options.instruction;
+    this.generativeSafetySettings = options.generativeSafetySettings;
     if (options.model) this.model = options.model;
 
     this.endpoint = `projects/${config.projectId}/locations/${config.location}/publishers/google/models/${this.model}`;
@@ -84,8 +76,9 @@ export class TextGenerator {
   }
 
   private initVertexClient() {
+    // here location is hard-coded, following https://cloud.google.com/vertex-ai/docs/generative-ai/embeddings/get-text-embeddings#generative-ai-get-text-embedding-nodejs
     const clientOptions = {
-      apiEndpoint: `${config.location}-prediction-aiplatform.googleapis.com`,
+      apiEndpoint: 'us-central1-aiplatform.googleapis.com',
     };
 
     this.vertexClient = new v1.PredictionServiceClient(clientOptions);
@@ -116,14 +109,7 @@ export class TextGenerator {
       const request = this.createVertexRequest(promptText, options);
       const [result] = await this.vertexClient.predict(request);
 
-      const {content, safetyAttributes} =
-        this.extractVertexCandidateResponse(result);
-
-      if (!content) {
-        return {candidates: [], safetyAttributes};
-      }
-
-      return {candidates: [content], safetyAttributes};
+      return this.extractVertexCandidateResponse(result);
     }
 
     if (!this.generativeClient) {
@@ -134,11 +120,7 @@ export class TextGenerator {
 
     const [result] = await this.generativeClient.generateText(request);
 
-    const candidates = this.extractGenerativeCandidationResponse(result);
-
-    return {
-      candidates,
-    };
+    return this.extractGenerativeCandidationResponse(result);
   }
 
   private createGenerativeRequest(
@@ -154,28 +136,15 @@ export class TextGenerator {
       topK: this.topK,
       temperature: this.temperature,
       candidateCount: this.candidateCount,
+      safetySettings: this.generativeSafetySettings,
       maxOutputTokens: this.maxOutputTokens,
       ...options,
     };
     return request;
   }
 
-  private extractGenerativeCandidationResponse(
-    result: APIGenerateTextResponse
-  ) {
-    if (!result.candidates || !result.candidates.length) {
-      throw new Error('No candidates returned from server.');
-    }
-
-    //TODO: do we need to filter out empty strings? This seems to be a type issue with the API, why are they optional?
-    const candidates = result.candidates
-      .map(candidate => candidate.output)
-      .filter(output => !!output) as string[];
-
-    if (!candidates.length) {
-      throw new Error('No candidates returned from server.');
-    }
-    return candidates;
+  private extractGenerativeCandidationResponse(result: GLGenerateTextResponse) {
+    return convertToTextGeneratorResponse(result as GenerativePrediction);
   }
 
   private createVertexRequest(
@@ -224,20 +193,70 @@ export class TextGenerator {
 
     const predictionValue = result.predictions[0] as protobuf.common.IValue;
 
-    const prediction = helpers.fromValue(predictionValue);
+    const vertexPrediction = helpers.fromValue(predictionValue);
 
-    const {safetyAttributes, content} = prediction as {
-      safetyAttributes?: {
-        blocked: boolean;
-        categories: string[];
-        scores: number[];
-      };
-      content?: string;
+    return convertToTextGeneratorResponse(vertexPrediction as VertexPrediction);
+  }
+}
+
+type VertexPrediction = {
+  safetyAttributes?: {
+    blocked: boolean;
+    categories: string[];
+    scores: number[];
+  };
+  content?: string;
+};
+
+type GenerativePrediction = {
+  candidates: {output: string}[];
+  filters?: {reason: string}[];
+  safetyFeedback?: {
+    rating: Record<string, any>;
+    setting: Record<string, any>;
+  }[];
+};
+
+type TextGeneratorResponse = {
+  candidates: string[];
+  safetyMetadata?: {
+    blocked: boolean;
+    [key: string]: any;
+  };
+};
+
+function convertToTextGeneratorResponse(
+  prediction: VertexPrediction | GenerativePrediction
+): TextGeneratorResponse {
+  // if it's generative language
+  if ('candidates' in prediction) {
+    const {candidates, filters, safetyFeedback} = prediction;
+    const blocked = !!filters && filters.length > 0;
+    const safetyMetadata = {
+      blocked,
+      safetyFeedback,
     };
-
+    if (!candidates.length && !blocked) {
+      throw new Error('No candidates returned from the Generative API.');
+    }
     return {
-      content,
+      candidates: candidates.map(candidate => candidate.output),
+      safetyMetadata,
+    };
+  } else {
+    // provider will be vertex
+    const {content, safetyAttributes} = prediction;
+    const blocked = !!safetyAttributes && !!safetyAttributes.blocked;
+    const safetyMetadata = {
+      blocked,
       safetyAttributes,
+    };
+    if (!content && !blocked) {
+      throw new Error('No content returned from the Vertex PaLM API.');
+    }
+    return {
+      candidates: blocked ? [] : [content!],
+      safetyMetadata,
     };
   }
 }