+++ firestore-palm-summarize-text/functions/src/index.ts
@@ -25,6 +25,8 @@ const {textField, responseField, collectionName, targetSummaryLength} = config;
 
 const textGenerator = new TextGenerator({
   model: config.model,
+  maxOutputTokens: config.maxOutputTokens,
+  generativeSafetySettings: config.generativeSafetySettings,
 });
 
 logs.init(config);
@@ -79,18 +81,26 @@ export const generateSummary = functions.firestore
         'status.updateTime': FieldValue.serverTimestamp(),
       };
 
-      if (result.safetyAttributes) {
-        metadata['safetyAttributes'] = result.safetyAttributes;
+      if (result.safetyMetadata) {
+        metadata.safetyMetadata = {};
+
+        /** Ensure only defined data is added to the metadata */
+        for (const key of Object.keys(result.safetyMetadata)) {
+          if (result.safetyMetadata[key] !== undefined) {
+            metadata.safetyMetadata[key] = result.safetyMetadata[key];
+          }
+        }
       }
 
-      if (result.safetyAttributes?.blocked) {
+      if (result.safetyMetadata?.blocked) {
         return ref.update({
           ...metadata,
           'status.state': 'ERRORED',
           'status.error':
-            'The text provided was blocked by the Vertex AI content filter.',
+            'The prompt or summary was blocked by the PaLM content filter.',
         });
       }
+
       return ref.update({
         ...metadata,
         [responseField]: result.candidates[0],
@@ -108,8 +118,9 @@ export const generateSummary = functions.firestore
   });
 
 const createSummaryPrompt = (text: string, targetSummaryLength?: number) => {
-  if (!targetSummaryLength) {
-    return `Summarize this text: "${text}"`;
-  }
-  return `Summarize this text in exactly ${targetSummaryLength} sentences: "${text}"`;
+  const prompt = `Give a summary of the following text in ${targetSummaryLength} sentences, do not use any information that is not explicitly mentioned in the text.
+  text: ${text}
+`;
+
+  return prompt;
 };