+++ firestore-semantic-search/functions/__tests__/functions/backfill_embeddings_task.test.ts
@@ -7,7 +7,7 @@ const mockUploadToCloudStorage = jest.fn();
 const mockSaveEmbeddingsToTmpFile = jest.fn();
 const mockDeleteTempFiles = jest.fn();
 
-jest.mock('utils', () => ({
+jest.mock('../../src/common/utils', () => ({
   saveEmbeddingsToTmpFile: (args: unknown) => mockSaveEmbeddingsToTmpFile(args),
   uploadToCloudStorage: (args: unknown) => mockUploadToCloudStorage(args),
   deleteTempFiles: (args: unknown) => mockDeleteTempFiles(args),
@@ -21,11 +21,11 @@ jest.mock('firebase-functions', () => ({
   },
 }));
 
-jest.mock('config', () => ({
+jest.mock('../../src/config', () => ({
   default: {
     // System vars
     location: 'us-central1',
-    projectId: 'dev-extensions-testing',
+    projectId: 'demo-gcp',
     instanceId: 'test-instance',
 
     // User-defined vars
@@ -38,7 +38,7 @@ jest.mock('config', () => ({
     tasksDoc: '_ext-test-instance/tasks',
     metadataDoc: '_ext-test-instance/metadata',
     dimensions: 512,
-    bucketName: 'dev-extensions-testing-ext-test-instance',
+    bucketName: 'demo-gcp-ext-test-instance',
   },
 }));