+++ storage-extract-image-text/functions/__tests__/startsWithArray.test.ts
@@ -3,7 +3,7 @@ import {startsWithArray} from '../src/util';
 describe('startsWithArray', () => {
   test('returns true when the image path starts with a path in the array', () => {
     const userInputPaths = ['/images'];
-    const imagePath = 'images/1234.jpg';
+    const imagePath = '/images/1234.jpg';
 
     const result = startsWithArray(userInputPaths, imagePath);