+++ storage-reverse-image-search/functions/src/common/vertex.ts
@@ -29,6 +29,8 @@ import {AlgorithmConfig} from '../types/algorithm_config';
 
 const apiEndpoint = `${config.location}-aiplatform.googleapis.com`;
 
+const AcceleratorType = protos.google.cloud.aiplatform.v1.AcceleratorType;
+
 export const indexClient = new aiplatform.IndexServiceClient({
   apiEndpoint: apiEndpoint,
   fallback: 'rest',
@@ -86,9 +88,9 @@ export async function createIndex(
               dimensions: {
                 numberValue: dimensions,
               },
-              // TODO make this configurable
               approximateNeighborsCount: {numberValue: config.neighbors},
               distanceMeasureType: {stringValue: config.distanceMeasureType},
+              shardSize: {stringValue: config.shardSize},
               featureNormType: {stringValue: config.featureNormType},
               algorithmConfig: {},
             },
@@ -140,11 +142,34 @@ export async function createIndexEndpoint() {
  * @param index format: projects/{project}/locations/{location}/indexes/{index}
  */
 export async function deployIndex(indexEndpoint: string, index: string) {
+  const acceleratorType =
+    AcceleratorType[config.acceleratorType as keyof typeof AcceleratorType];
+
+  // If acceleratorType is unspecified, acceleratorCount must be undefined.
+  const acceleratorCount =
+    acceleratorType !== AcceleratorType.ACCELERATOR_TYPE_UNSPECIFIED
+      ? config.acceleratorCount
+      : undefined;
+
   const [operation] = await indexEndpointClient.deployIndex({
     indexEndpoint: indexEndpoint,
     deployedIndex: {
       id: `ext_${config.instanceId.replace(/-/g, '_')}_index`,
       index: index,
+      dedicatedResources: {
+        /** DedicatedResources machineSpec */
+        machineSpec: {
+          machineType: config.machineType,
+          acceleratorType: acceleratorType,
+          acceleratorCount: acceleratorCount,
+        },
+
+        /** DedicatedResources minReplicaCount */
+        minReplicaCount: config.minReplicaCount,
+
+        /** DedicatedResources maxReplicaCount */
+        maxReplicaCount: config.maxReplicaCount,
+      },
     },
   });
 
@@ -320,5 +345,6 @@ export async function checkIndexStatus(): Promise<{
   indexEndpoint?: string;
 }> {
   const metdata = await admin.firestore().doc(config.metadataDoc).get();
+
   return metdata.data() ?? {};
 }