+++ firestore-palm-gen-text/CHANGELOG.md
@@ -1,6 +1,17 @@
+## Version 0.1.8
+
+- Added parameters for configuring filter thresholds on developers PaLM API
+- Increased Max content length default
+## Version 0.1.7
+
+- Fixed location bug where apiEndpoint was incorrectly entered
+
 ## Version 0.1.6
 
-Removed unnecessary Storage bucket requirement from PREINSTALL.
+- Removed unnecessary Storage bucket requirement from PREINSTALL.
+- Support both Vertex and Generative Language PaLM API Providers.
+- Handle PaLM Content Filters.
+- Add support for providing an API key for the Generative Language PaLM Provider.
 
 ## Version 0.1.5